+++ src/gui/Blog/Blog.Service/Domain/Model/EntryTopic.cs
@@ -2,6 +2,6 @@
 
 public class EntryTopic
 {
-    public uint? Id { get; set; }
+    public long? Id { get; set; }
     public string? Name { get; set; }
 }
